requirements.txt
torch
torchtext
lightning

jupyter
numpy
pandas
matplotlib

python-dotenv
icecream
boto3

transformers
accelerate
evaluate
bitsandbytes
datasets
peft
scikit-learn

streamlit
huggingface_hub